[fix] address comments

+++ paimon-core/src/main/java/org/apache/paimon/operation/SnapshotDeletion.java
@@ -29,7 +29,7 @@ import org.apache.paimon.manifest.ManifestFile;
 import org.apache.paimon.manifest.ManifestFileMeta;
 import org.apache.paimon.manifest.ManifestList;
 import org.apache.paimon.utils.FileStorePathFactory;
-import org.apache.paimon.utils.Triple;
+import org.apache.paimon.utils.Pair;
 
 import org.apache.paimon.shade.guava30.com.google.common.collect.Iterables;
 
@@ -78,17 +78,17 @@ public class SnapshotDeletion {
      *
      * @param manifestListName name of manifest list
      * @param deletionBuckets partition-buckets of which some data files have been deleted
-     * @param skippingTester a tester that if the test result of a data file's manifest entry is
-     *     true, the data file will be skipped
+     * @param dataFileSkipper if the test result of a data file is true, the data file will be
+     *     skipped when deleting
      */
     public void deleteExpiredDataFiles(
             String manifestListName,
             Map<BinaryRow, Set<Integer>> deletionBuckets,
-            Predicate<ManifestEntry> skippingTester) {
+            Predicate<TagFileKeeper.DataFileInfo> dataFileSkipper) {
         doDeleteExpiredDataFiles(
                 getManifestEntriesFromManifestList(manifestListName),
                 deletionBuckets,
-                skippingTester);
+                dataFileSkipper);
     }
 
     /**
@@ -155,11 +155,9 @@ public class SnapshotDeletion {
      * file.
      *
      * @param skipped manifest file deletion skipping set, deleted manifest file will be added to
-     *     this set too.
-     * @param deleteDataManifestList whether to delete the manifest list files of data.
+     *     this set too. NOTE: changelog manifests won't be checked.
      */
-    public void deleteManifestFiles(
-            Set<ManifestFileMeta> skipped, Snapshot snapshot, boolean deleteDataManifestList) {
+    public void deleteManifestFiles(Set<String> skipped, Snapshot snapshot) {
         // cannot call `toExpire.dataManifests` directly, it is possible that a job is
         // killed during expiration, so some manifest files may have been deleted
         List<ManifestFileMeta> toExpireManifests = new ArrayList<>();
@@ -168,9 +166,10 @@ public class SnapshotDeletion {
 
         // delete manifest
         for (ManifestFileMeta manifest : toExpireManifests) {
-            if (!skipped.contains(manifest)) {
-                manifestFile.delete(manifest.fileName());
-                skipped.add(manifest);
+            String fileName = manifest.fileName();
+            if (!skipped.contains(fileName)) {
+                manifestFile.delete(fileName);
+                skipped.add(fileName);
             }
         }
         if (snapshot.changelogManifestList() != null) {
@@ -181,8 +180,10 @@ public class SnapshotDeletion {
         }
 
         // delete manifest lists
-        if (deleteDataManifestList) {
+        if (!skipped.contains(snapshot.baseManifestList())) {
             manifestList.delete(snapshot.baseManifestList());
+        }
+        if (!skipped.contains(snapshot.deltaManifestList())) {
             manifestList.delete(snapshot.deltaManifestList());
         }
         if (snapshot.changelogManifestList() != null) {
@@ -205,11 +206,11 @@ public class SnapshotDeletion {
     void doDeleteExpiredDataFiles(
             Iterable<ManifestEntry> dataFileLog,
             Map<BinaryRow, Set<Integer>> deletionBuckets,
-            Predicate<ManifestEntry> skippingTester) {
+            Predicate<TagFileKeeper.DataFileInfo> dataFileSkipper) {
         // we cannot delete a data file directly when we meet a DELETE entry, because that
         // file might be upgraded
-        // data file path -> (partition, bucket, extra file paths)
-        Map<Path, Triple<BinaryRow, Integer, List<Path>>> dataFileToDelete = new HashMap<>();
+        // data file path -> (data file info, extra file paths)
+        Map<Path, Pair<TagFileKeeper.DataFileInfo, List<Path>>> dataFileToDelete = new HashMap<>();
         for (ManifestEntry entry : dataFileLog) {
             Path bucketPath = pathFactory.bucketPath(entry.partition(), entry.bucket());
             Path dataFilePath = new Path(bucketPath, entry.file().fileName());
@@ -218,17 +219,13 @@ public class SnapshotDeletion {
                     dataFileToDelete.remove(dataFilePath);
                     break;
                 case DELETE:
-                    // check whether we should skip the data file
-                    if (skippingTester.test(entry)) {
-                        break;
-                    }
-
                     List<Path> extraFiles = new ArrayList<>(entry.file().extraFiles().size());
                     for (String file : entry.file().extraFiles()) {
                         extraFiles.add(new Path(bucketPath, file));
                     }
                     dataFileToDelete.put(
-                            dataFilePath, Triple.of(entry.partition(), entry.bucket(), extraFiles));
+                            dataFilePath,
+                            Pair.of(TagFileKeeper.DataFileInfo.of(entry), extraFiles));
                     break;
                 default:
                     throw new UnsupportedOperationException(
@@ -237,12 +234,18 @@ public class SnapshotDeletion {
         }
 
         dataFileToDelete.forEach(
-                (path, triple) -> {
-                    // delete data files
-                    fileIO.deleteQuietly(path);
-                    triple.f2.forEach(fileIO::deleteQuietly);
-                    // record changed buckets
-                    deletionBuckets.computeIfAbsent(triple.f0, p -> new HashSet<>()).add(triple.f1);
+                (path, pair) -> {
+                    TagFileKeeper.DataFileInfo info = pair.getLeft();
+                    // check whether we should skip the data file
+                    if (!dataFileSkipper.test(info)) {
+                        // delete data files
+                        fileIO.deleteQuietly(path);
+                        pair.getRight().forEach(fileIO::deleteQuietly);
+                        // record changed buckets
+                        deletionBuckets
+                                .computeIfAbsent(info.partition, p -> new HashSet<>())
+                                .add(info.bucket);
+                    }
                 });
     }
